refactor: usePlanetNavigate hooks
